refactor(components): migrate Dropdown to atomic design structure
- Move Dropdown component to atomic design directory structure
- Update import paths in Header and User components to match new location
- Maintains existing functionality while following project's atomic design principles
