+++ src/container/Header/components/User.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from '@emotion/styled'
 import { breakpoint } from '@styles/utils/breakpoint'
 import { useState } from 'react'
-import Dropdown from '@components/Dropdown/Dropdown'
+import Dropdown from '@components/atoms/Dropdown/Dropdown'
 import UserMenu from '@container/Header/components/dropdown/UserMenu'
 import { useSelector } from 'react-redux'
 import { logout } from '@store/userSlice'
